something odd was happening, and not causing a successful logon form submission. Either it is mechanize or the YG site.  Either way, this hidden field 'mydb_events[210]' wasn't putting the value in the HTML from in the Post.  It was cutting it off.  To get around this, because that particular element is hidden, I had to unhide them all.
I tried to just unhide the one, but it wasn't working.  So they all get
unhidden  and I can force it to the value the html is giving.  Would
definitely like to not do this, but not possible right now.  Took a bit
of time to debug, but it is working now.